requirements.txt
transformers
rich
speedtest-cli
func_timeout
PyYAML
